refactor: extract ResourceTitleBar component
